+++ public/service/resources.go
@@ -2,6 +2,7 @@ package service
 
 import (
 	"context"
+	"time"
 
 	"github.com/benthosdev/benthos/v4/internal/bundle"
 	"github.com/benthosdev/benthos/v4/internal/component/cache"
@@ -19,14 +20,35 @@ func newResourcesFromManager(nm bundle.NewManagement) *Resources {
 }
 
 // MockResources returns an instantiation of a resources struct that provides
-// valid but ineffective methods and observability components. This is useful
-// for testing components that interact with a resources type but do not
-// explicitly need it for testing purposes.
-func MockResources() *Resources {
-	// This is quite naughty, if we encounter a case where an empty resource
-	// config like this could actually return an error then we'd need to change
-	// this.
-	return newResourcesFromManager(mock.NewManager())
+// valid but ineffective methods and observability components. It is possible to
+// instantiate this with mocked (in-memory) cache and rate limit types for
+// testing purposed.
+func MockResources(opts ...MockResourcesOptFn) *Resources {
+	m := mock.NewManager()
+	for _, o := range opts {
+		o(m)
+	}
+	return newResourcesFromManager(m)
+}
+
+// MockResourcesOptFn provides a func based optional argument to MockResources.
+type MockResourcesOptFn func(*mock.Manager)
+
+// MockResourcesOptAddCache instantiates the resources type with a mock cache
+// with a given name. Cached items are held in memory.
+func MockResourcesOptAddCache(name string) MockResourcesOptFn {
+	return func(m *mock.Manager) {
+		m.Caches[name] = map[string]mock.CacheItem{}
+	}
+}
+
+// MockResourcesOptAddRateLimit instantiates the resources type with a mock rate
+// limit with a given name, the provided closure will be called for each
+// invocation of the rate limit.
+func MockResourcesOptAddRateLimit(name string, fn func(context.Context) (time.Duration, error)) MockResourcesOptFn {
+	return func(m *mock.Manager) {
+		m.RateLimits[name] = mock.RateLimit(fn)
+	}
 }
 
 // Label returns a label that identifies the component instantiation. This could